+++ miden/src/main.rs
@@ -14,7 +14,12 @@ pub(crate) mod utils;
 
 /// Root CLI struct
 #[derive(Parser, Debug)]
-#[command(name = "Miden", about = "Miden CLI", version, rename_all = "kebab-case")]
+#[command(
+    name = "miden-vm",
+    about = "The Miden virtual machine",
+    version,
+    rename_all = "kebab-case"
+)]
 pub struct Cli {
     #[command(subcommand)]
     action: Actions,