+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 #### Changes
 
+- [BREAKING] Rename `miden` executable to `miden-vm`
 - Improve error messages for some assembler instruction (#1785)
 - Remove `idx` column from Kernel ROM chiplet and use chiplet bus for initialization. (#1818)
 - [BREAKING] Make `Assembler::source_manager()` be `Send + Sync` (#1822)